improve setup and nvim

+++ setup.sh
@@ -3,9 +3,6 @@
 function setup_ubuntu() {
     sudo apt update && sudo apt upgrade -y
 
-    sudo add-apt-repository ppa:neovim-ppa/stable -y
-    sudo apt update
-
     sudo apt install -y \
         neovim \
         build-essential \
@@ -18,7 +15,12 @@ function setup_ubuntu() {
 
     chsh -s $(which zsh)
 
-    curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.0/install.sh | bash
+    tmpdir=$(mktemp -d)
+    wget -O $tmpdir/nvim-linux64.deb https://github.com/neovim/neovim/releases/download/stable/nvim-linux64.deb
+    sudo apt install $tmpdir/nvim-linux64.deb
+    rm -rf $tmpdir
+
+    curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.0/install.sh | zsh
 
 
 }
@@ -52,7 +54,10 @@ function setup() {
     done
 
     mkdir -p ~/.config
-    ln -s $dotdir/nvim ~/.config/nvim
+    if [[ ! -L ~/.config/nvim ]]
+    then
+        ln -s $dotdir/nvim ~/.config/nvim
+    fi
 
     if [[ ! -d ~/.tmux/plugins/tpm ]] 
     then
